asdfasdf

+++ cnn.py
@@ -94,7 +94,7 @@ class ConvLayer(object):
                 log.debug("self.z_values[j][i].shape : %s", type(self.z_values[j][i]))
 
                 # self.z_values[j][i] = np.add(np.sum(np.multiply(input_neurons[:, row:self.filter_size + row, slide:self.filter_size + slide], self.weights[j])), self.biases[j])
-                self.z_values[j][i] = np.sum(input_neurons[:, row:self.filter_size + row, slide:self.filter_size + slide] * self.weights[j]) + self.biases[j]
+                self.z_values[j][i] = float(np.sum(input_neurons[:, row:self.filter_size + row, slide:self.filter_size + slide] * self.weights[j]) + self.biases[j])
 
                 # self.z_values[j][i] = f
                 log.debug("self.z_values[j][i].shape : %s", self.z_values[j][i])
